pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "parallax"
version = "0.1.0"
description = "Parallax: Zero-Shot Controllable Image-to-Video via Test-Time Guidance with Vision Foundation Models"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "diffusers>=0.32.0",
    "transformers>=4.47.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "einops>=0.7.0",
]

[project.optional-dependencies]
sam2 = ["segment-anything-2>=0.1.0"]
dev = ["pytest>=7.4.0", "pytest-mock>=3.12.0"]
all = ["parallax[sam2,dev]"]

[tool.setuptools.packages.find]
where = ["src"]
